refactor: split debug_cmds.py into debug/ package
Split the 692-line debug_cmds.py into a well-organized package:

**Package structure:**
- `debug/__init__.py` - Public API (10 cmd_* functions)
- `debug/_helpers.py` - Internal: _build_probe
- `debug/openocd_cmds.py` - OpenOCD server management (4 commands)
- `debug/gdb_cmds.py` - GDB batch/script execution (2 commands)
- `debug/inspection_cmds.py` - Variable/memory inspection (4 commands)

**Grouping logic:**
1. **OpenOCD** - Server lifecycle (status, start, stop, cmd)
2. **GDB** - Script execution (gdb, gdb_script)
3. **Inspection** - High-level debugging (inspect, threads, watch, memdump)

**Test updates:**
- Updated imports from `eab.cli.debug_cmds` → `eab.cli.debug`
- Fixed patches to point to correct module namespaces
- Updated var_cmds.py to import _build_probe from debug._helpers

**Results:**
- ✅ All 12 debug tests passing
- ✅ Clean separation by functionality
- ✅ Public API preserved
- ✅ No breaking changes
